# api/docs/swagger.yaml
definitions:
...
  models.CarbonComparison:
    properties:
      bpan_a:
        type: string
      bpan_a_lower:
        type: boolean
      bpan_b:
        type: string
      emissions_per_kwh_delta:
        type: number
      stage1_delta:
        type: number
      stage2_delta:
        type: number
      stage3_delta:
        type: number
      stage4_delta:
        type: number
      stage5_delta:
        type: number
      total_delta:
        type: number
    type: object
  models.CarbonFootprintRequest:
    properties:
      cell_production_method:
        type: string
      factory_energy_source:
        type: string
      manufacturing_emissions_kg_co2e:
        type: number
      manufacturing_location:
        type: string
      mining_method:
        type: string
      raw_material_emissions_kg_co2e:
        type: number
      raw_material_source_country:
        type: string
      recycling_avoided_mining:
        type: number
      recycling_emissions_kg_co2e:
        type: number
      recycling_method:
        type: string
      recycling_recovery_rate:
        type: number
      transport_distance_km:
        type: number
      transport_emissions_kg_co2e:
        type: number
      transport_mode:
        type: string
      transport_packaging:
        type: string
      usage_annual_km:
        type: integer
      usage_emissions_kg_co2e:
        type: number
      usage_grid_emissions_factor_gco2e_per_kwh:
        type: number
      usage_years:
        type: integer
    type: object
  models.CarbonFootprintResponse:
    properties:
      bpan:
        type: string
      emissions_per_kwh:
        type: number
      total_emissions_kg_co2e:
        type: number
      verification_standard:
        type: string
      verified:
        type: boolean
      verified_at:
        type: string
      verified_by:
        type: string
    type: object
...
  models.VerificationRequest:
    properties:
      standard:
        description: ISO 14040, PEF, EU ETS
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/batteries/{bpan_a}/carbon/compare/{bpan_b}:
    get:
      consumes:
      - application/json
      description: Emission delta per stage, identify lower-emission battery
      parameters:
      - description: First BPAN
        in: path
        name: bpan_a
        required: true
        type: string
      - description: Second BPAN
        in: path
        name: bpan_b
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Comparison result
          schema:
            $ref: '#/definitions/models.CarbonComparison'
      summary: Compare carbon footprints (A vs B)
      tags:
      - carbon
  /api/v1/batteries/{bpan}/carbon:
    get:
      consumes:
      - application/json
      description: Retrieve BCF (all roles can view verified data)
      parameters:
      - description: BPAN
        in: path
        name: bpan
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Carbon footprint
          schema:
            $ref: '#/definitions/models.CarbonFootprintResponse'
      summary: Get battery carbon footprint
      tags:
      - carbon
    post:
      consumes:
      - application/json
      description: Submit BCF with raw material, manufacturing, transport, usage,
        recycling emissions
      parameters:
      - description: BPAN
        in: path
        name: bpan
        required: true
        type: string
      - description: Carbon data (5 stages)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CarbonFootprintRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Submission ID
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Unauthorized (non-manufacturer)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Submit battery carbon footprint (5-stage emissions)
      tags:
      - carbon
  /api/v1/batteries/{bpan}/carbon/verify:
    post:
      consumes:
      - application/json
      description: Mark carbon data as verified against standard (ISO 14040, PEF,
        EU ETS)
      parameters:
      - description: BPAN
        in: path
        name: bpan
        required: true
        type: string
      - description: Verification details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.VerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Verified
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Unauthorized (non-verifier)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Verify carbon footprint (third-party verifier only)
      tags:
      - carbon
...
swagger: "2.0"